app/backend/requirements.txt
Django==5.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
requests==2.31.0
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
